gs_conv/requirements.txt
wandb==0.16.3
hydra-core==1.2.0
einops==0.7.0
torch==2.2.0
pytest==7.4.0
torchvision==0.17.0